Finish basic layout
