+++ backend/package.json
@@ -7,11 +7,9 @@
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
 	"dependencies": {
-		"@discordjs/collection": "^1.5.3",
-		"dotenv": "^16.3.1",
 		"express": "^4.18.2",
 		"mongoose": "^7.5.3",
-		"openai": "^4.11.0"
+		"openai": "^4.11.1"
 	},
 	"devDependencies": {
 		"@types/express": "^4.17.18",
