+++ common/ChatBot.ts
@@ -1,0 +1,9 @@
+export class ChatBot {
+	public name: string;
+	public prompt: string;
+
+	constructor(name: string, prompt: string) {
+		this.name = name;
+		this.prompt = prompt;
+	}
+}
