+++ frontend/index.html
@@ -7,7 +7,7 @@
 		<link rel="stylesheet" href="./src/output.css" />
 		<title>AI Personalities React App</title>
 	</head>
-	<body class="bg-slate-50 dark:bg-slate-600 text-black dark:text-white">
+	<body>
 		<div id="root"></div>
 		<script type="module" src="/src/main.tsx"></script>
 	</body>
