+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import SideBar from './components/SideBar';
 
 export default function App() {
 	return (
-		<div className="App">
+		<div className="App bg-slate-50 dark:bg-slate-700 text-black dark:text-white min-w-screen w-screen h-full">
 			<SideBar />
 		</div>
 	);
