+++ frontend/tailwind.config.ts
@@ -1,6 +1,10 @@
 import { type Config } from 'tailwindcss';
 
 export default {
-	content: ['./src/**/*.{html,tsx}'],
+	content: [
+		'./src/*.{html,tsx}',
+		'./src/**/*.{html,tsx}',
+		'./src/**/**/*.{html,tsx}'
+	],
 	darkMode: 'class'
 } satisfies Config;
