+++ package.json
@@ -8,7 +8,11 @@
 		"eslint-plugin-react-hooks": "^4.6.0",
 		"eslint-plugin-react-refresh": "^0.4.3",
 		"npm-check-updates": "^16.14.4",
+		"openai": "^4.11.1",
 		"prettier": "^3.0.3",
 		"typescript": "^5.2.2"
+	},
+	"dependencies": {
+		"@discordjs/collection": "^1.5.3"
 	}
 }
